LPS-111561 Isolate pagination components
