+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/item-configuration-panels/collection-general-panel/CollectionGeneralPanel.js
@@ -12,14 +12,7 @@
  * details.
  */
 
-import ClayForm, {
-	ClayCheckbox,
-	ClayInput,
-} from '@clayui/form';
-import ClayIcon from '@clayui/icon';
 import {useModal} from '@clayui/modal';
-import {useIsMounted} from '@liferay/frontend-js-react-web';
-import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React, {useCallback, useEffect, useMemo, useState} from 'react';
 
@@ -44,27 +37,18 @@ import {getResponsiveConfig} from '../../../../../../../app/utils/getResponsiveC
 import {useId} from '../../../../../../../app/utils/useId';
 import Collapse from '../../../../../../../common/components/Collapse';
 import CollectionSelector from '../../../../../../../common/components/CollectionSelector';
-import useControlledState from '../../../../../../../core/hooks/useControlledState';
 import CollectionFilterConfigurationModal from '../../CollectionFilterConfigurationModal';
 import {CommonStyles} from '../CommonStyles';
 import {LayoutSelector} from './LayoutSelector';
 import {ListItemStyleSelector} from './ListItemStyleSelector';
+import {NoPaginationOptions} from './NoPaginationOptions';
+import {PaginationOptions} from './PaginationOptions';
 import {PaginationSelector} from './PaginationSelector';
 import {ShowGutterSelector} from './ShowGutterSelector';
 import {StyleDisplaySelector} from './StyleDisplaySelector';
 import {VerticalAlignmentSelector} from './VerticalAlignmentSelector';
 
 const LIST_STYLE_GRID = '';
-
-const ERROR_MESSAGES = {
-	maximumItems: Liferay.Language.get(
-		'the-current-number-of-items-in-this-collection-is-x'
-	),
-	maximumItemsPerPage: Liferay.Language.get(
-		'you-can-only-display-a-maximum-of-x-items-per-page'
-	),
-	noItems: Liferay.Language.get('this-collection-has-no-items'),
-};
 
 export function CollectionGeneralPanel({item}) {
 	const {
@@ -89,10 +73,7 @@ export function CollectionGeneralPanel({item}) {
 	const collectionPaginationTypeId = useId();
 	const dispatch = useDispatch();
 	const getState = useGetState();
-	const isMounted = useIsMounted();
 	const segmentsExperienceId = useSelector(selectSegmentsExperienceId);
-
-	const [totalNumberOfItems, setTotalNumberOfItems] = useState(0);
 
 	const {
 		observer: filterConfigurationObserver,
@@ -195,19 +176,6 @@ export function CollectionGeneralPanel({item}) {
 		},
 		[item.itemId, dispatch, segmentsExperienceId, selectedViewportSize]
 	);
-
-	useEffect(() => {
-		if (collection) {
-			CollectionService.getCollectionItemCount({
-				collection,
-				onNetworkStatus: () => {},
-			}).then(({totalNumberOfItems}) => {
-				if (isMounted()) {
-					setTotalNumberOfItems(totalNumberOfItems);
-				}
-			});
-		}
-	}, [collection, isMounted]);
 
 	useEffect(() => {
 		if (collection && listStyle && listStyle !== LIST_STYLE_GRID) {
@@ -364,15 +332,13 @@ export function CollectionGeneralPanel({item}) {
 										/>
 									) : (
 										<NoPaginationOptions
+											collection={collection}
 											displayAllItems={displayAllItems}
 											handleConfigurationChanged={
 												handleConfigurationChanged
 											}
 											initialNumberOfItems={
 												initialNumberOfItems
-											}
-											totalNumberOfItems={
-												totalNumberOfItems
 											}
 										/>
 									)}
@@ -404,302 +370,4 @@ export function CollectionGeneralPanel({item}) {
 
 CollectionGeneralPanel.propTypes = {
 	item: PropTypes.object.isRequired,
-};
-
-function NoPaginationOptions({
-	displayAllItems,
-	handleConfigurationChanged,
-	initialNumberOfItems,
-	totalNumberOfItems,
-}) {
-	const collectionNumberOfItemsId = useId();
-
-	const [numberOfItems, setNumberOfItems] = useControlledState(
-		initialNumberOfItems
-	);
-	const [numberOfItemsError, setNumberOfItemsError] = useState(null);
-
-	useEffect(() => {
-		let errorMessage = null;
-
-		if (totalNumberOfItems) {
-			if (initialNumberOfItems > totalNumberOfItems) {
-				errorMessage = Liferay.Util.sub(
-					ERROR_MESSAGES.maximumItems,
-					totalNumberOfItems
-				);
-			}
-		}
-		else {
-			errorMessage = ERROR_MESSAGES.noItems;
-		}
-
-		setNumberOfItemsError(errorMessage);
-	}, [totalNumberOfItems, initialNumberOfItems]);
-
-	const handleDisplayAllItemsChanged = (event) =>
-		handleConfigurationChanged({
-			displayAllItems: event.target.checked,
-		});
-
-	const handleCollectionNumberOfItemsBlurred = (event) => {
-		const nextValue = Math.abs(Number(event.target.value)) || 1;
-
-		if (!numberOfItems || numberOfItems < 0) {
-			setNumberOfItems(nextValue);
-		}
-
-		if (nextValue !== initialNumberOfItems) {
-			handleConfigurationChanged({
-				numberOfItems: nextValue,
-			});
-		}
-	};
-
-	return (
-		<>
-			<div className="mb-2 pt-1">
-				<ClayCheckbox
-					checked={displayAllItems}
-					label={Liferay.Language.get('display-all-collection-items')}
-					onChange={handleDisplayAllItemsChanged}
-				/>
-			</div>
-
-			{displayAllItems && (
-				<p className="mt-1 small text-secondary">
-					{Liferay.Util.sub(
-						Liferay.Language.get(
-							'this-setting-can-affect-page-performance-severely-if-the-number-of-collection-items-is-above-x.-we-strongly-recommend-using-pagination-instead'
-						),
-						config.searchContainerPageMaxDelta
-					)}
-				</p>
-			)}
-
-			{!displayAllItems && (
-				<ClayForm.Group
-					className={classNames({
-						'has-warning': numberOfItemsError,
-					})}
-					small
-				>
-					<label htmlFor={collectionNumberOfItemsId}>
-						{Liferay.Language.get(
-							'maximum-number-of-items-to-display'
-						)}
-					</label>
-
-					<ClayInput
-						id={collectionNumberOfItemsId}
-						min="1"
-						onBlur={handleCollectionNumberOfItemsBlurred}
-						onChange={(event) =>
-							setNumberOfItems(Number(event.target.value))
-						}
-						type="number"
-						value={numberOfItems || ''}
-					/>
-
-					<p className="mt-1 small text-secondary">
-						{Liferay.Util.sub(
-							Liferay.Language.get(
-								'setting-a-value-above-x-can-affect-page-performance-severely'
-							),
-							config.searchContainerPageMaxDelta
-						)}
-					</p>
-
-					{numberOfItemsError && (
-						<FeedbackMessage message={numberOfItemsError} />
-					)}
-				</ClayForm.Group>
-			)}
-		</>
-	);
-}
-
-NoPaginationOptions.propTypes = {
-	displayAllItems: PropTypes.bool.isRequired,
-	handleConfigurationChanged: PropTypes.func.isRequired,
-	initialNumberOfItems: PropTypes.number.isRequired,
-	totalNumberOfItems: PropTypes.number.isRequired,
-};
-
-function PaginationOptions({
-	displayAllPages,
-	handleConfigurationChanged,
-	initialNumberOfItemsPerPage,
-	initialNumberOfPages,
-}) {
-	const collectionNumberOfItemsPerPageId = useId();
-	const collectionNumberOfPagesId = useId();
-
-	const [numberOfItemsPerPage, setNumberOfItemsPerPage] = useControlledState(
-		initialNumberOfItemsPerPage
-	);
-	const [numberOfItemsPerPageError, setNumberOfItemsPerPageError] = useState(
-		null
-	);
-	const [numberOfPages, setNumberOfPages] = useState(initialNumberOfPages);
-
-	const isMaximumValuePerPageError =
-		initialNumberOfItemsPerPage > config.searchContainerPageMaxDelta;
-
-	useEffect(() => {
-		let errorMessage = null;
-
-		if (isMaximumValuePerPageError) {
-			errorMessage = Liferay.Util.sub(
-				ERROR_MESSAGES.maximumItemsPerPage,
-				config.searchContainerPageMaxDelta
-			);
-		}
-		else if (initialNumberOfItemsPerPage < 1) {
-			errorMessage = ERROR_MESSAGES.neededItem;
-		}
-
-		setNumberOfItemsPerPageError(errorMessage);
-	}, [isMaximumValuePerPageError, initialNumberOfItemsPerPage]);
-
-	const handleCollectionNumberOfItemsPerPageBlurred = (event) => {
-		const nextValue = Math.abs(Number(event.target.value)) || 1;
-
-		if (!numberOfItemsPerPage || numberOfItemsPerPage < 0) {
-			setNumberOfItemsPerPage(nextValue);
-		}
-
-		if (nextValue !== initialNumberOfItemsPerPage) {
-			handleConfigurationChanged({
-				numberOfItemsPerPage: nextValue,
-			});
-		}
-	};
-
-	const handleCollectionNumberOfPagesBlurred = (event) => {
-		const nextValue = Math.abs(Number(event.target.value)) || 1;
-
-		if (!numberOfPages || numberOfPages < 0) {
-			setNumberOfPages(nextValue);
-		}
-
-		if (nextValue !== initialNumberOfPages) {
-			handleConfigurationChanged({
-				numberOfPages: nextValue,
-			});
-		}
-	};
-
-	const handleDisplayAllPagesChanged = (event) =>
-		handleConfigurationChanged({
-			displayAllPages: event.target.checked,
-		});
-
-	return (
-		<>
-			<div className="mb-2 pt-1">
-				<ClayCheckbox
-					checked={displayAllPages}
-					label={Liferay.Language.get('display-all-pages')}
-					onChange={handleDisplayAllPagesChanged}
-				/>
-			</div>
-
-			{!displayAllPages && (
-				<ClayForm.Group small>
-					<label htmlFor={collectionNumberOfPagesId}>
-						{Liferay.Language.get(
-							'maximum-number-of-pages-to-display'
-						)}
-					</label>
-
-					<ClayInput
-						id={collectionNumberOfPagesId}
-						min="1"
-						onBlur={handleCollectionNumberOfPagesBlurred}
-						onChange={(event) =>
-							setNumberOfPages(Number(event.target.value))
-						}
-						type="number"
-						value={numberOfPages || ''}
-					/>
-				</ClayForm.Group>
-			)}
-
-			<ClayForm.Group
-				className={classNames({
-					'has-warning': numberOfItemsPerPageError,
-				})}
-				small
-			>
-				<label htmlFor={collectionNumberOfItemsPerPageId}>
-					{Liferay.Language.get('maximum-number-of-items-per-page')}
-				</label>
-
-				<ClayInput
-					id={collectionNumberOfItemsPerPageId}
-					min="1"
-					onBlur={handleCollectionNumberOfItemsPerPageBlurred}
-					onChange={(event) =>
-						setNumberOfItemsPerPage(Number(event.target.value))
-					}
-					type="number"
-					value={numberOfItemsPerPage || ''}
-				/>
-
-				<div className="mb-2 mt-1">
-					<span
-						className={classNames(
-							'mr-1 small',
-							isMaximumValuePerPageError &&
-								numberOfItemsPerPageError
-								? 'text-warning'
-								: 'text-secondary',
-							{
-								'd-none':
-									isMaximumValuePerPageError &&
-									numberOfItemsPerPageError,
-							}
-						)}
-					>
-						{Liferay.Util.sub(
-							Liferay.Language.get('x-items-maximum'),
-							config.searchContainerPageMaxDelta
-						)}
-					</span>
-
-					{numberOfItemsPerPageError && (
-						<FeedbackMessage message={numberOfItemsPerPageError} />
-					)}
-				</div>
-			</ClayForm.Group>
-		</>
-	);
-}
-
-PaginationOptions.propTypes = {
-	displayAllPages: PropTypes.bool.isRequired,
-	handleConfigurationChanged: PropTypes.func.isRequired,
-	initialNumberOfItemsPerPage: PropTypes.number.isRequired,
-	initialNumberOfPages: PropTypes.number.isRequired,
-};
-
-function FeedbackMessage({message}) {
-	return (
-		<div className="autofit-row mt-2 small text-warning">
-			<div className="autofit-col">
-				<div className="autofit-section mr-2">
-					<ClayIcon symbol="warning-full" />
-				</div>
-			</div>
-
-			<div className="autofit-col autofit-col-expand">
-				<div className="autofit-section">{message}</div>
-			</div>
-		</div>
-	);
-}
-
-FeedbackMessage.propTypes = {
-	message: PropTypes.string.isRequired,
 };
